Let clients cache images
The URLs dispensed to the user include a timestamp for the time the file was
last modified. This is necessary, because LBG supports editing images
(rotation/ resizing, etc) and it does this without changing the filepath the
user sees. But by including the timestamp as a query param, we're forcing the
browser to check for an updated copy.