fix: Retain caption when resizing or rotating jpeg images
The limitation with how images in a gallery are tracked right now means
we need to update the meta records from the old filename to the new one
anytime we perform actions that result in a "new" file being added to the
pool.

Without this, the caption on uploaded images that aren't jpg or png lose
the caption when being edited.